Optimize in-memory schema storage

+++ src/main/java/com/zendesk/maxwell/schema/columndef/BigIntColumnDef.java
@@ -7,7 +7,7 @@ public class BigIntColumnDef extends ColumnDef {
 
 	protected boolean signed;
 
-	public BigIntColumnDef(String name, String type, int pos, boolean signed) {
+	public BigIntColumnDef(String name, String type, short pos, boolean signed) {
 		super(name, type, pos);
 		this.signed = signed;
 	}
