+++ src/main/java/com/zendesk/maxwell/schema/columndef/BitColumnDef.java
@@ -4,7 +4,7 @@ import java.math.BigInteger;
 import java.util.BitSet;
 
 public class BitColumnDef extends ColumnDef {
-	public BitColumnDef(String name, String type, int pos) {
+	public BitColumnDef(String name, String type, short pos) {
 		super(name, type, pos);
 	}
